ショートカットキーのmenubarを修正
