+++ lib/main.dart
@@ -15,7 +15,6 @@ import 'package:quick_flutter/firebase_options.dart';
 import 'package:quick_flutter/model/analytics_event/analytics_event_name.dart';
 import 'package:quick_flutter/screen/settings/settings_screen.dart';
 import 'package:quick_flutter/screen/todo_screen/todo_screen.dart';
-import 'package:quick_flutter/controller/url/url_controller.dart';
 import 'package:quick_flutter/widget/actions/delete_todo/delete_todo_action.dart';
 import 'package:quick_flutter/widget/actions/focus_down/focus_down_action.dart';
 import 'package:quick_flutter/widget/actions/focus_up/focus_up_action.dart';
@@ -28,6 +27,7 @@ import 'package:quick_flutter/widget/actions/select_todo_down/select_todo_down_a
 import 'package:quick_flutter/widget/actions/select_todo_up/select_todo_up_action.dart';
 import 'package:quick_flutter/widget/actions/toggle_focus/toggle_focus_action.dart';
 import 'package:quick_flutter/widget/actions/toggle_todo_done/toggle_todo_done_action.dart';
+import 'package:quick_flutter/widget/app_platform_menu_bar.dart';
 import 'package:quick_flutter/widget/shortcutkey.dart';
 import 'package:quick_flutter/widget/theme/app_theme.dart';
 
@@ -44,7 +44,7 @@ void main() async {
       observers: [_AppObserver()],
       child: AppMaterialApp(
         home: _CallbackShortcuts(
-          child: _PlatformMenuBar(
+          child: AppPlatformMenuBar(
             child: Consumer(
               builder: (context, ref, child) {
                 final largeWindowKey = GlobalKey();
@@ -203,179 +203,6 @@ class _LoggingActionDispatcher extends ActionDispatcher {
     }
 
     return super.invokeActionIfEnabled(action, intent, context);
-  }
-}
-
-/// [PlatformMenuBar]の中でRefを使うためにラップしたWidgetクラス
-class _PlatformMenuBar extends ConsumerWidget {
-  const _PlatformMenuBar({required this.child});
-  final Widget child;
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final channel = ref.watch(methodChannelProvider);
-    return PlatformMenuBar(
-      menus: [
-        PlatformMenu(
-          label: "",
-          menus: [
-            PlatformMenuItemGroup(
-              members: [
-                PlatformMenuItem(
-                  label: 'Quit Tokeru',
-                  shortcut: ShortcutActivatorType.quit.shortcutActivator,
-                  onSelected: () {
-                    channel.invokeMethod(
-                      AppMethodChannel.quit.name,
-                    );
-                  },
-                ),
-              ],
-            ),
-          ],
-        ),
-
-        // todo
-        PlatformMenu(
-          label: "Todo",
-          menus: [
-            PlatformMenuItemGroup(
-              members: [
-                // 新規TODO
-                PlatformMenuItem(
-                  label: ShortcutActivatorType.newTodo.label,
-                  shortcut: ShortcutActivatorType.newTodo.shortcutActivator,
-                  onSelected: () {
-                    Actions.maybeInvoke<NewTodoIntent>(
-                      context,
-                      const NewTodoIntent(),
-                    );
-                  },
-                ),
-                // フォーカス中のTODOをチェックする
-                PlatformMenuItem(
-                  label: ShortcutActivatorType.toggleDone.label,
-                  shortcut: ShortcutActivatorType.toggleDone.shortcutActivator,
-                  onSelected: () {
-                    Actions.maybeInvoke<ToggleTodoDoneIntent>(
-                      context,
-                      const ToggleTodoDoneIntent(),
-                    );
-                  },
-                ),
-                // Todoを削除
-                PlatformMenuItem(
-                  label: ShortcutActivatorType.deleteTodo.label,
-                  shortcut: ShortcutActivatorType.deleteTodo.shortcutActivator,
-                  onSelected: () {
-                    Actions.maybeInvoke<DeleteTodoIntent>(
-                      context,
-                      const DeleteTodoIntent(),
-                    );
-                  },
-                ),
-              ],
-            ),
-            PlatformMenuItemGroup(
-              members: [
-                // 上へ移動
-                PlatformMenuItem(
-                  label: ShortcutActivatorType.moveUp.label,
-                  shortcut: ShortcutActivatorType.moveUp.shortcutActivator,
-                  onSelected: () => Actions.maybeInvoke<MoveUpTodoIntent>(
-                    context,
-                    const MoveUpTodoIntent(),
-                  ),
-                ),
-                // 下へ移動
-                PlatformMenuItem(
-                  label: ShortcutActivatorType.moveDown.label,
-                  shortcut: ShortcutActivatorType.moveDown.shortcutActivator,
-                  onSelected: () {
-                    Actions.maybeInvoke<MoveDownTodoIntent>(
-                      context,
-                      const MoveDownTodoIntent(),
-                    );
-                  },
-                ),
-              ],
-            ),
-          ],
-        ),
-
-        // window
-        PlatformMenu(
-          label: "Window",
-          menus: [
-            PlatformMenuItemGroup(
-              members: [
-                PlatformMenuItem(
-                  label: ShortcutActivatorType.reload.label,
-                  onSelected: () {
-                    Actions.maybeInvoke<ReloadIntent>(
-                      context,
-                      const ReloadIntent(),
-                    );
-                  },
-                  shortcut: ShortcutActivatorType.reload.shortcutActivator,
-                ),
-              ],
-            ),
-          ],
-        ),
-
-        // contact developer
-        PlatformMenu(
-          label: "You are welcome to contact me!👋",
-          menus: [
-            PlatformMenuItemGroup(
-              members: [
-                PlatformMenuItem(
-                  label: 'Thank you for using Tokeru!😊',
-                  onSelected: () async {
-                    await UrlController.developerXAccount.launch();
-                  },
-                ),
-                PlatformMenuItem(
-                  label: 'I would like to hear your feedback!',
-                  onSelected: () async {
-                    await UrlController.developerXAccount.launch();
-                  },
-                ),
-              ],
-            ),
-            PlatformMenuItemGroup(
-              members: [
-                PlatformMenuItem(
-                  label: '📩 Follow on X(Twitter)',
-                  onSelected: () async {
-                    await UrlController.developerXAccount.launch();
-                  },
-                ),
-                PlatformMenuItem(
-                  label: '💡 Got an idea for a feature',
-                  onSelected: () async {
-                    await UrlController.featureRequest.launch();
-                  },
-                ),
-                PlatformMenuItem(
-                  label: '📝 Found a bug',
-                  onSelected: () async {
-                    await UrlController.bugReport.launch();
-                  },
-                ),
-                PlatformMenuItem(
-                  label: '🧑‍💻 Tokeru repository is public',
-                  onSelected: () async {
-                    await UrlController.tokeruRepository.launch();
-                  },
-                ),
-              ],
-            ),
-          ],
-        ),
-      ],
-      child: child,
-    );
   }
 }
